Доработка тестов, доработка апдейта заданий
